verification of product key complete. endpoints protected.

+++ src/main/java/com/chemcalc/practice/endpoints/CompoundEndpoint.java
@@ -6,6 +6,7 @@ import java.util.Random;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -13,15 +14,22 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
+import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.chemcalc.practice.domain.Compound;
+import com.chemcalc.practice.domain.ProgramKey;
 import com.chemcalc.practice.controller.CompoundService;
+import com.chemcalc.practice.controller.ProgramKeyService;
 
 @RestController
 public class CompoundEndpoint {
 	@Autowired
 	CompoundService compoundService;
+	
+	@Autowired
+	ProgramKeyService programKeyService;
 	
 	@GetMapping("allCompounds")
 	public Iterable<Compound> allCompounds() {
@@ -34,22 +42,28 @@ public class CompoundEndpoint {
 	}
 	
 	@PostMapping("addCompound")
-	public void addCompound(@RequestBody Compound compound) {
-		compoundService.saveCompound(compound);
-	}
-	
-	@PutMapping("madeUpHeader")
-	public void header(@RequestHeader("made-up") String thing) { // This works, just make sure to add the header to the CORS filter
-		System.out.println(thing);
-		//include API key in headers for DB-altering endpoints
-		//And check if the provided key is in the database
-		//The 'login' is also just a key-check to let a user onto the page
+	public void addCompound(@RequestHeader("PRODUCT_KEY") String key, @RequestBody Compound compound) {
+		//Standard HTTP error codes suffice here. 
+		//Throwing a ResponseStatusException will get the code across
+		//The front-end checks for this code, handling the exception if it comes up for clear user feedback.
+		//This is an alternative to sending a whole ResponseEntity as a return object.
+		Optional<ProgramKey> keyOptional = programKeyService.findByKeyString(key);
+		if (keyOptional.isEmpty()) {
+			throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "No valid product key found");
+		} else {
+			compoundService.saveCompound(compound);			
+		}
 	}
 	
 	@DeleteMapping("deleteCompound")
-	public void deleteCompound(@RequestBody String recievedId) {
-		long id = Long.parseLong(recievedId);
-		compoundService.deleteCompound(id);
+	public void deleteCompound(@RequestHeader("PRODUCT_KEY") String key, @RequestBody String recievedId) {
+		Optional<ProgramKey> keyOptional = programKeyService.findByKeyString(key);
+		if (keyOptional.isEmpty()) {
+			throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "No valid product key found");
+		} else {
+			long id = Long.parseLong(recievedId);
+			compoundService.deleteCompound(id);						
+		}
 	}
 	
 	@GetMapping("randomCompound/{limit}")
@@ -62,8 +76,13 @@ public class CompoundEndpoint {
 	
 	//UpdateCompound is the only untested endpoint rn
 	@PutMapping("updateCompound")
-	public void updateCompound(@RequestBody Compound compound) {
-		//Note: no id supplied in url. Make sure it's in the compound object!
+	public void updateCompound(@RequestHeader("PRODUCT_KEY") String key, @RequestBody Compound compound) {
+		Optional<ProgramKey> keyOptional = programKeyService.findByKeyString(key);
+		if (keyOptional.isEmpty()) {
+			throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "No valid product key found");
+		} 
+		
+		//Note: no id supplied in url. Make sure it's in the compound object on the client-side.
 		Optional<Compound> compoundOptional = compoundService.findById(compound.getId());
 		
 		if (compoundOptional.isEmpty()) {
